fix: 修复了一系列问题，彻底支持了 author_id
