+++ src/plugins/github/plugins/remove/depends.py
@@ -1,30 +1,14 @@
-from githubkit.rest import (
-    PullRequestPropLabelsItems,
-    WebhookIssueCommentCreatedPropIssueAllof0PropLabelsItems,
-    WebhookIssuesEditedPropIssuePropLabelsItems,
-    WebhookIssuesOpenedPropIssuePropLabelsItems,
-    WebhookIssuesReopenedPropIssueMergedLabels,
-    WebhookPullRequestReviewSubmittedPropPullRequestPropLabelsItems,
-)
-from githubkit.typing import Missing
-
 from nonebot.params import Depends
 
 from src.plugins.github.depends import get_labels
+from src.plugins.github.typing import LabelsItems
 
 
 def get_name_by_labels(
-    labels: list[PullRequestPropLabelsItems]
-    | list[WebhookPullRequestReviewSubmittedPropPullRequestPropLabelsItems]
-    | Missing[list[WebhookIssuesOpenedPropIssuePropLabelsItems]]
-    | Missing[list[WebhookIssuesReopenedPropIssueMergedLabels]]
-    | Missing[list[WebhookIssuesEditedPropIssuePropLabelsItems]]
-    | list[WebhookIssueCommentCreatedPropIssueAllof0PropLabelsItems] = Depends(
-        get_labels
-    ),
+    labels: LabelsItems = Depends(get_labels),
 ) -> list[str]:
     """通过标签获取名称"""
-    label_names = []
+    label_names: list[str] = []
     if not labels:
         return label_names
 
@@ -34,7 +18,7 @@ def get_name_by_labels(
     return label_names
 
 
-def check_labels(labels: list[str] | str):  # -> Any:
+def check_labels(labels: list[str] | str):
     """检查标签是否存在"""
     if isinstance(labels, str):
         labels = [labels]
@@ -42,9 +26,6 @@ def check_labels(labels: list[str] | str):  # -> Any:
     async def _check_labels(
         has_labels: list[str] = Depends(get_name_by_labels),
     ) -> bool:
-        for label in labels:
-            if label not in has_labels:
-                return False
-        return True
+        return all(label in has_labels for label in labels)
 
     return Depends(_check_labels)
