+++ src/plugins/github/typing.py
@@ -1,4 +1,4 @@
-from typing import TypeAlias
+from typing import TypeAlias, TypedDict
 from nonebot.adapters.github import (
     IssueCommentCreated,
     IssuesEdited,
@@ -8,8 +8,32 @@ from nonebot.adapters.github import (
     PullRequestReviewSubmitted,
 )
 
-# 暂不用 type 关键字，编辑器不支持
+from githubkit.rest import (
+    PullRequestPropLabelsItems,
+    WebhookIssueCommentCreatedPropIssueAllof0PropLabelsItems,
+    WebhookIssuesEditedPropIssuePropLabelsItems,
+    WebhookIssuesOpenedPropIssuePropLabelsItems,
+    WebhookIssuesReopenedPropIssuePropLabelsItems,
+    WebhookPullRequestReviewSubmittedPropPullRequestPropLabelsItems,
+)
+from githubkit.typing import Missing
+
 IssuesEvent: TypeAlias = (
     IssuesOpened | IssuesReopened | IssuesEdited | IssueCommentCreated
 )
+
 PullRequestEvent: TypeAlias = PullRequestClosed | PullRequestReviewSubmitted
+
+LabelsItems: TypeAlias = (
+    list[PullRequestPropLabelsItems]
+    | list[WebhookPullRequestReviewSubmittedPropPullRequestPropLabelsItems]
+    | Missing[list[WebhookIssuesOpenedPropIssuePropLabelsItems]]
+    | Missing[list[WebhookIssuesReopenedPropIssuePropLabelsItems]]
+    | Missing[list[WebhookIssuesEditedPropIssuePropLabelsItems]]
+    | list[WebhookIssueCommentCreatedPropIssueAllof0PropLabelsItems]
+)
+
+
+class AuthorInfo(TypedDict):
+    author: str
+    author_id: int
