+++ src/plugins/github/utils.py
@@ -1,10 +1,15 @@
 import json
+import subprocess
+
 from pathlib import Path
 from re import Pattern
-import subprocess
 from typing import Any
 from nonebot import logger
+
 from pydantic_core import to_jsonable_python
+from githubkit.rest import Issue
+
+from src.plugins.github.typing import AuthorInfo
 
 
 def run_shell_command(command: list[str]):
@@ -55,3 +60,13 @@ def extract_publish_info_from_issue(
         for key, match in matchers.items()
     }
     return data
+
+
+def extract_author_info(issue: Issue) -> AuthorInfo:
+    """
+    从议题中获取作者信息
+    """
+    return {
+        "author": issue.user.login if issue.user else "",
+        "author_id": issue.user.id if issue.user else 0,
+    }
